# lib/api-zod/src/generated/types/alertExplanation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface AlertExplanation {
  alertId: number;
  explanation: string;
  recommendations: string;
  /** @nullable */
  sources?: string | null;
}