# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface DashboardSummary {
  totalRecords: number;
  totalPrescriptions: number;
  activeGrants: number;
  unresolvedAlerts: number;
  criticalAlerts: number;
  verifiedRecords?: number;
  recentActivity: ActivityItem[];
}