# lib/api-zod/src/generated/types/healthRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthRecord {
  id: number;
  title: string;
  category: string;
  encryptedContent: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  recordDate?: string | null;
  signatureHash: string;
  isVerified: boolean;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}