# lib/api-zod/src/generated/types/healthRecordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthRecordInput {
  /** @minLength 1 */
  title: string;
  category: string;
  content: string;
  provider?: string;
  recordDate?: string;
  tags?: string;
}