# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  dateOfBirth: string;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  encryptionKeyFingerprint: string;
  /** @nullable */
  publicKey?: string | null;
  createdAt: string;
}