# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface Prescription {
  id: number;
  medicationName: string;
  dosage: string;
  frequency: string;
  prescribedBy: string;
  prescribedAt: string;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  refillsRemaining?: number | null;
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  pharmacyName?: string | null;
  /** @nullable */
  npiNumber?: string | null;
  status: string;
  signatureHash: string;
  createdAt: string;
}