# lib/api-zod/src/generated/types/signatureVerification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediVault Patient Digital Health Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface SignatureVerification {
  recordId: number;
  isValid: boolean;
  signatureHash: string;
  algorithm?: string;
  verifiedAt: string;
  /** @nullable */
  message?: string | null;
}